elastic2csv/requirements.txt
astroid==2.9.3
certifi==2021.10.8
-e git+ssh://git@github.com/bernardcosta/elastic2csv.git@8de8f6635896a00ab4aad552c0dae024aa3683c1#egg=elastic2csv&subdirectory=elastic2csv
elasticsearch==7.16.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
numpy==1.21.5
pandas==1.3.5
platformdirs==2.5.0
progressbar==2.5
pylint==2.12.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
toml==0.10.2
tqdm==4.62.3
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
